zeromq/requirements.txt
python-dotenv==1.0.0
pyzmq==25.1.1